+++ R/SDistribution_WeightedDiscrete.R
@@ -314,6 +314,12 @@ WeightedDiscrete <- R6Class("WeightedDiscrete",
       pdf <- self$getParameterValue("pdf")
 
       if (checkmate::testList(data)) {
+        # hacky fix for uneven vectors
+        lng <- min(sapply(data, length))
+        for (i in seq_along(pdf)) {
+          pdf[[i]] <- pdf[[i]][seq.int(lng)]
+          data[[i]] <- data[[i]][seq.int(lng)]
+        }
         pdf <- matrix(unlist(pdf), nrow = length(data[[1]]), ncol = length(data))
         data <- matrix(unlist(data), ncol = ncol(pdf))
         return(C_Vec_WeightedDiscretePdf(x, data, pdf, log))
@@ -344,6 +350,12 @@ WeightedDiscrete <- R6Class("WeightedDiscrete",
       cdf <- self$getParameterValue("cdf")
 
       if (checkmate::testList(data)) {
+        # hacky fix for uneven vectors
+        lng <- min(sapply(data, length))
+        for (i in seq_along(cdf)) {
+          cdf[[i]] <- cdf[[i]][seq.int(lng)]
+          data[[i]] <- data[[i]][seq.int(lng)]
+        }
         cdf <- matrix(unlist(cdf), nrow = length(data[[1]]), ncol = length(data))
         data <- matrix(unlist(data), ncol = ncol(cdf))
         return(C_Vec_WeightedDiscreteQuantile(p, data, cdf, lower.tail, log.p))